FIX annotate plugin: round numbers for json output
The calculation in get.php can produce numbers with fractions.
When using the German locale, PHP uses a ',' as the decimal separator.
This leads to invalid JSON. As far as I understand, the fraction is
not actually needed. Therefore, the numbers are rounded before printing
them as JSON. If the fractions are indeed needed, the numbers should be
formatted using something like number_format().

git-svn-id: http://svn.montala.net/svn/resourcespace@5347 c08608d7-6e46-0410-86ca-f2a6f1370df5